Move AllocationParadigm to dmod.core.
Moving AllocationParadigm enum class from dmod.scheduler.job to
dmod.core, under new execution.py module, to better detach it for use in
various places it is required, without also requiring the jobs package.
