+++ python/lib/scheduler/dmod/scheduler/job/__init__.py
@@ -1,3 +1,3 @@
-from .job import Job, AllocationParadigm, JobExecPhase, JobExecStep, JobImpl, JobStatus, RequestedJob
+from .job import Job, JobExecPhase, JobExecStep, JobImpl, JobStatus, RequestedJob
 from .job_util import JobUtil, DefaultJobUtilFactory
 from .job_manager import JobManager, JobManagerFactory
